Show version on restart and add version command
