+++ src/commands/fetch/fetch.rs
@@ -1,4 +1,3 @@
-use crate::extensions::*;
 use anyhow::*;
 
 use super::*;
