+++ src/events/mod.rs
@@ -19,32 +19,17 @@ mod message;
 mod message_delete;
 mod message_update;
 mod reaction_add;
+mod ready;
 
 pub struct Handler;
 
 #[async_trait]
 impl EventHandler for Handler {
-    async fn ready(&self, ctx: client::Context, _data_about_bot: Ready) {
-        log::info!("Trup is ready!");
-        {
-            let config = ctx.get_config().await;
-
-            match load_up_emotes(&ctx, config.guild).await {
-                Ok(emotes) => {
-                    ctx.data.write().await.insert::<UPEmotes>(Arc::new(emotes));
-                }
-                Err(err) => {
-                    log::warn!("Error loading emotes: {}", err);
-                }
-            }
-        }
-
-        let _ = ctx
-            .set_presence(Some(Activity::listening("!help")), OnlineStatus::Online)
-            .await;
-
-        start_mute_handler(ctx.clone()).await;
-        start_attachment_log_handler(ctx).await;
+    async fn ready(&self, ctx: client::Context, data_about_bot: Ready) {
+        log_error!(
+            "Error while handling ready event",
+            ready::ready(ctx, data_about_bot).await
+        )
     }
 
     async fn message(&self, ctx: client::Context, msg: Message) {
@@ -143,70 +128,4 @@ async fn unmute(
     member.remove_roles(&ctx, &[config.role_mute]).await?;
 
     Ok(())
-}
-
-async fn start_mute_handler(ctx: client::Context) {
-    tokio::spawn(async move {
-        let (config, db) = ctx.get_config_and_db().await;
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-            let mutes = match db.get_newly_expired_mutes().await {
-                Ok(mutes) => mutes,
-                Err(err) => {
-                    log::error!("Failed to request expired mutes: {}", err);
-                    continue;
-                }
-            };
-            for mute in mutes {
-                if let Err(err) = unmute(&ctx, &config, &db, &mute).await {
-                    log::error!("Error handling mute removal: {}", err);
-                } else {
-                    config
-                        .log_bot_action(&ctx, |e| {
-                            e.description(format!("{} is now unmuted", mute.user.mention()));
-                        })
-                        .await;
-                }
-            }
-        }
-    });
-}
-
-async fn start_attachment_log_handler(ctx: client::Context) {
-    tokio::spawn(async move {
-        let config = ctx.get_config().await;
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(10)).await;
-
-            log_error!(
-                "Failed to clean up attachments",
-                crate::attachment_logging::cleanup(&config).await
-            );
-        }
-    });
-}
-
-async fn load_up_emotes(ctx: &client::Context, guild: GuildId) -> Result<UPEmotes> {
-    let all_emoji = guild.emojis(&ctx).await?;
-    Ok(UPEmotes {
-        pensibe: all_emoji
-            .iter()
-            .find(|x| x.name == "pensibe")
-            .context("no pensibe emote found")?
-            .clone(),
-        police: all_emoji
-            .iter()
-            .find(|x| x.name == "police")
-            .context("no police emote found")?
-            .clone(),
-        poggers: all_emoji
-            .iter()
-            .find(|x| x.name == "poggersphisch")
-            .context("no police poggers found")?
-            .clone(),
-        stares: all_emoji
-            .into_iter()
-            .filter(|x| x.name.starts_with("stare"))
-            .collect(),
-    })
 }
